requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
feedparser>=6.0.0
newspaper3k>=0.2.8
